// src/commons/connector/yaci/types/drep-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Yaci Store API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { DRepRegistration } from "./drep-registration";
/**
 *
 *
 * @export
 * @interface DRepDto
 */
export interface DRepDto {
  /**
   * @type {string}
   * @memberof DRepDto
   */
  drepHash?: string;

  /**
   * @type {string}
   * @memberof DRepDto
   */
  drepId?: string;

  /**
   * @type {Array<DRepRegistration>}
   * @memberof DRepDto
   */
  registrations?: Array<DRepRegistration>;

  /**
   * @type {string}
   * @memberof DRepDto
   */
  anchorUrl?: string;

  /**
   * @type {string}
   * @memberof DRepDto
   */
  anchorHash?: string;

  /**
   * @type {number}
   * @memberof DRepDto
   */
  createdAt?: number;

  /**
   * @type {number}
   * @memberof DRepDto
   */
  updatedAt?: number;
}